feat: updated audio message ui
