+++ src/features/rooms/components/message/gif-message.tsx
@@ -1,0 +1,7 @@
+interface GifMessageProps {
+  url: string;
+}
+
+export default function GifMessage({ url }: GifMessageProps) {
+  return <video src={url} autoPlay loop muted playsInline />;
+}
